backend/requirements.txt
fastapi
uvicorn
python-dotenv
botbuilder-core
gunicorn
aiohttp
openai>=1.0.0